Move enterprise service to a separate file
